refactor：TSC access reconfiguration
