+++ src/Infrastructure/Masa.Mc.Infrastructure.Tsc/Masa.Mc.Infrastructure.Tsc.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Masa.Contrib.StackSdks.Tsc" Version="0.6.0-preview.3" />
+  </ItemGroup>
+
+</Project>
