+++ src/Services/Masa.Mc.Service/_Imports.cs
@@ -156,12 +156,6 @@ global using Masa.BuildingBlocks.ReadWriteSpliting.Cqrs.Queries;
 global using Masa.Mc.Contracts.Admin.Dtos.MessageTasks.Imports;
 global using Magicodes.ExporterAndImporter.Core.Models;
 global using Masa.Utils.Caching.Core.Interfaces;
-global using Masa.Contrib.StackSdks.Tsc;
-global using OpenTelemetry.Logs;
-global using OpenTelemetry.Metrics;
-global using OpenTelemetry.Resources;
-global using OpenTelemetry.Trace;
-global using Masa.Mc.Service.Admin.Infrastructure.Extensions;
 global using Masa.Mc.Infrastructure.MailKit;
 global using Masa.BuildingBlocks.StackSdks.Scheduler;
 global using Masa.BuildingBlocks.StackSdks.Scheduler.Model;
@@ -177,3 +171,4 @@ global using Microsoft.Extensions.Diagnostics.HealthChecks;
 global using Masa.BuildingBlocks.Authentication.Identity;
 global using Masa.Contrib.StackSdks.Mc.Infrastructure.Extensions;
 global using Masa.BuildingBlocks.StackSdks.Mc;
+global using Masa.Mc.Infrastructure.Tsc;
