+++ src/Services/Masa.Mc.Service/appsettings.Development.json
@@ -41,5 +41,11 @@
   "Environment": "Development",
   "ConfigObjects": [ "AppSettings", "AutoComplete", "RedisConfig" ], //待挂载的对象名, 此处会将Redis配置挂载到ConfigurationAPI:<Replace-With-Your-AppId>节点下
   "Secret": "", //Dcc App 秘钥
-  "Cluster": "Default"
+  "Cluster": "Default",
+  "MasaTscOption": {
+    "ServiceName": "masa-mc-service",
+    "ServiceNameSpace": "Development",
+    "ServiceVersion": "1.0.0"
+  },
+  "OtlpUrl": "http://otel.lonsid.cn:32247"
 }
