+++ src/Services/Masa.Mc.Service/appsettings.Staging.json
@@ -32,5 +32,11 @@
   "Environment": "Staging",
   "ConfigObjects": [ "AppSettings", "AutoComplete", "RedisConfig" ],
   "Secret": "",
-  "Cluster": "Default"
+  "Cluster": "Default",
+  "MasaTscOption": {
+    "ServiceName": "masa-mc-service",
+    "ServiceNameSpace": "Staging",
+    "ServiceVersion": "1.0.0"
+  },
+  "OtlpUrl": "http://otel-collector-dev.open-telemetry:4317"
 }
