+++ src/Services/Masa.Mc.Service/appsettings.json
@@ -34,9 +34,10 @@
   "AuthClient": {
     "Url": "https://auth-service-develop.masastack.com/"
   },
-  "OTLP": {
-    "Endpoint": "http://10.10.90.114:32147",
+  "MasaTscOption": {
     "ServiceName": "masa-mc-service",
+    "ServiceNameSpace": "Development",
     "ServiceVersion": "1.0.0"
-  }
+  },
+  "OtlpUrl": "http://otel.lonsid.cn:32247"
 }
