+++ src/Web/Masa.Mc.Web.Admin.Server/Program.cs
@@ -2,6 +2,8 @@
 // Licensed under the Apache License. See LICENSE.txt in the project root for license information.
 
 var builder = WebApplication.CreateBuilder(args);
+
+builder.AddObservability();
 // Add services to the container.
 builder.Services.AddRazorPages();
 builder.Services.AddServerSideBlazor();
