+++ src/Web/Masa.Mc.Web.Admin.Server/appsettings.Develop.json
@@ -12,6 +12,12 @@
     "Authority": "https://sso-develop.masastack.com",
     "ClientId": "masa.stack.web-develop",
     "ClientSecret": ""
-  }
+  },
+  "MasaTscOption": {
+    "ServiceName": "masa-mc-web",
+    "ServiceNameSpace": "Develop",
+    "ServiceVersion": "1.0.0"
+  },
+  "OtlpUrl": "http://otel-collector-dev.open-telemetry:4317"
 }
