+++ src/Web/Masa.Mc.Web.Admin.Server/appsettings.Development.json
@@ -13,5 +13,11 @@
     "Authority": "https://sso-develop.masastack.com",
     "ClientId": "masa.stack.web-development",
     "ClientSecret": ""
-  }
+  },
+  "MasaTscOption": {
+    "ServiceName": "masa-mc-web",
+    "ServiceNameSpace": "Development",
+    "ServiceVersion": "1.0.0"
+  },
+  "OtlpUrl": "http://otel.lonsid.cn:32247"
 }
