+++ src/Web/Masa.Mc.Web.Admin.Server/appsettings.Staging.json
@@ -12,6 +12,12 @@
     "Authority": "https://sso-staging.masastack.com",
     "ClientId": "masa.stack.web-staging",
     "ClientSecret": ""
-  }
+  },
+  "MasaTscOption": {
+    "ServiceName": "masa-mc-web",
+    "ServiceNameSpace": "Staging",
+    "ServiceVersion": "1.0.0"
+  },
+  "OtlpUrl": "http://otel-collector-dev.open-telemetry:4317"
 }
