+++ src/Web/Masa.Mc.Web.Admin.Server/appsettings.json
@@ -6,5 +6,11 @@
     }
   },
   "AllowedHosts": "*",
-  "McServiceBaseAddress": ""
+  "McServiceBaseAddress": "",
+  "MasaTscOption": {
+    "ServiceName": "masa-mc-web",
+    "ServiceNameSpace": "Development",
+    "ServiceVersion": "1.0.0"
+  },
+  "OtlpUrl": "http://otel.lonsid.cn:32247"
 }
